login_data/consumer/requirements.txt
paho-mqtt
diskcache
requests